Add/Delete track form database

+++ app/src/main/java/com/kuzheevadel/vmplayerv2/viewmodels/PlaybackViewModel.kt
@@ -5,12 +5,12 @@ import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
 import android.net.Uri
 import android.util.Log
+import com.kuzheevadel.vmplayerv2.common.DataBaseInfo
 import com.kuzheevadel.vmplayerv2.common.Source
 import com.kuzheevadel.vmplayerv2.common.UpdateUIMessage
 import com.kuzheevadel.vmplayerv2.database.PlaylistDatabase
 import com.kuzheevadel.vmplayerv2.database.TrackDao
 import com.kuzheevadel.vmplayerv2.interfaces.Interfaces
-import com.kuzheevadel.vmplayerv2.model.Track
 import com.kuzheevadel.vmplayerv2.repository.RadioRepository
 import io.reactivex.Completable
 import io.reactivex.schedulers.Schedulers
@@ -21,10 +21,13 @@ import javax.inject.Inject
 
 class PlaybackViewModel @Inject constructor(private val mediaRepository: Interfaces.StorageMediaRepository,
                                             private val radioRepository: RadioRepository,
-                                            private val database: PlaylistDatabase): ViewModel() {
+                                            database: PlaylistDatabase): ViewModel() {
 
     val trackData: MutableLiveData<UpdateUIMessage> = MutableLiveData()
     private val trackDao: TrackDao = database.trackDao()
+    val dataBaseInfoData: MutableLiveData<DataBaseInfo> = MutableLiveData()
+    val checkPlaylistData: MutableLiveData<DataBaseInfo> = MutableLiveData()
+
     lateinit var source: Source
 
     fun initViewModel() {
@@ -35,7 +38,7 @@ class PlaybackViewModel @Inject constructor(private val mediaRepository: Interfa
 
                     with(track) {
                         trackData.value =
-                            UpdateUIMessage(title, artist, albumId, null, duration, albumName, Source.TRACK)
+                            UpdateUIMessage(title, artist, albumId, null, duration, albumName, Source.TRACK, id, track.inPlaylist)
                     }
                 }
                 Source.RADIO -> {
@@ -47,7 +50,7 @@ class PlaybackViewModel @Inject constructor(private val mediaRepository: Interfa
                         Uri.parse(radioStation.favicon),
                         0,
                         "",
-                        Source.RADIO)
+                        Source.RADIO, -1, false)
                 }
             }
         } catch (e: Exception) {
@@ -57,30 +60,54 @@ class PlaybackViewModel @Inject constructor(private val mediaRepository: Interfa
         EventBus.getDefault().register(this)
     }
 
+    @SuppressLint("CheckResult")
     @Subscribe(threadMode = ThreadMode.MAIN)
     fun updateUI(message: UpdateUIMessage) {
         trackData.value = message
+
+        if (message.inPlaylist) {
+            checkPlaylistData.value = DataBaseInfo.ADDED
+        } else {
+            checkPlaylistData.value = DataBaseInfo.DONT_ADDED
+        }
     }
 
     @SuppressLint("CheckResult")
-    fun addTrackToPlaylistDatabase() {
-        Completable.fromAction { trackDao.insertTrack(mediaRepository.getCurrentTrack()) }
-            .subscribeOn(Schedulers.io())
-            .subscribe(
-                {
-                    Log.i("InsertTest", "in onNext")
-                    EventBus.getDefault().post("post")
-                },
-                {
+    fun addOrDeleteTrackFromPlaylist() {
 
-                }
-            )
-    }
+        if (!mediaRepository.getCurrentTrack().inPlaylist) {
+            mediaRepository.getCurrentTrack().inPlaylist = true
 
-    fun deleteTrackFromPlaylistDatabase() {
-        Completable.fromAction { trackDao.deleteTrack(mediaRepository.getCurrentTrack()) }
-            .subscribeOn(Schedulers.io())
-            .subscribe()
+            Completable.fromAction { trackDao.insertTrack(mediaRepository.getCurrentTrack()) }
+                .subscribeOn(Schedulers.io())
+                .subscribe(
+                    {
+                        Log.i("InsertTest", "in onNext")
+                        EventBus.getDefault().post("post")
+                        mediaRepository.setFlagById(mediaRepository.getCurrentTrack().id, true)
+                        dataBaseInfoData.postValue(DataBaseInfo.ADDED)
+                    },
+                    {
+                        Log.e("InsertTest", "Add error", it)
+                        dataBaseInfoData.postValue(DataBaseInfo.ERROR)
+                    }
+                )
+        } else {
+            mediaRepository.getCurrentTrack().inPlaylist = false
+
+            Completable.fromAction { trackDao.deleteTrack(mediaRepository.getCurrentTrack()) }
+                .subscribeOn(Schedulers.io())
+                .subscribe(
+                    {
+                        EventBus.getDefault().post("post")
+                        mediaRepository.setFlagById(mediaRepository.getCurrentTrack().id, false)
+                        dataBaseInfoData.postValue(DataBaseInfo.DELETED)
+                    },
+                    {
+                        dataBaseInfoData.postValue(DataBaseInfo.ERROR)
+                    }
+                )
+        }
     }
 
     override fun onCleared() {
